removed unnecessary function and interface
