+++ oci/registry.go
@@ -78,16 +78,6 @@ func WithRemoteOptions(opts ...remote.Option) Option {
 
 func (o *RegistryOptions) ClientOpts(ctx context.Context) ([]Option, error) {
 	opts := []Option{WithRemoteOptions(o.GetRegistryClientOpts(ctx)...)}
-	// if o.RefOpts.TagPrefix != "" {
-	// 	opts = append(opts, ociremote.WithPrefix(o.RefOpts.TagPrefix))
-	// }
-	// targetRepoOverride, err := ociremote.GetEnvTargetRepository()
-	// if err != nil {
-	// 	return nil, err
-	// }
-	// if (targetRepoOverride != name.Repository{}) {
-	// 	opts = append(opts, ociremote.WithTargetRepository(targetRepoOverride))
-	// }
 	return opts, nil
 }
